+++ src/frontmenu_ingame_map.c
@@ -294,8 +294,8 @@ int draw_overlay_call_to_arms(struct PlayerInfo *player, long units_per_px, long
                 // Now rotate the coordinates to receive minimap points
                 long mapos_x;
                 long mapos_y;
-                mapos_x = (zmpos_x * LbCosL(interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
-                mapos_y = (zmpos_y * LbCosL(interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
+                mapos_x = (zmpos_x * LbCosL((long)interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
+                mapos_y = (zmpos_y * LbCosL((long)interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
                 draw_call_to_arms_circle(thing->owner, 0, 0, mapos_x, mapos_y, zoom);
                 n++;
             }
@@ -431,8 +431,8 @@ int draw_overlay_spells_and_boxes(struct PlayerInfo *player, long units_per_px, 
                 long mapos_x;
                 long mapos_y;
                 // Now rotate the coordinates to receive minimap points
-                mapos_x = (zmpos_x * LbCosL(interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
-                mapos_y = (zmpos_y * LbCosL(interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
+                mapos_x = (zmpos_x * LbCosL((long)interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
+                mapos_y = (zmpos_y * LbCosL((long)interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
                 RealScreenCoord basepos;
                 basepos = MapDiagonalLength/2;
 
@@ -570,8 +570,8 @@ int draw_overlay_creatures(struct PlayerInfo *player, long units_per_px, long zo
                 // Now rotate the coordinates to receive minimap points
                 long mapos_x;
                 long mapos_y;
-                mapos_x = (zmpos_x * LbCosL(interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
-                mapos_y = (zmpos_y * LbCosL(interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
+                mapos_x = (zmpos_x * LbCosL((long)interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
+                mapos_y = (zmpos_y * LbCosL((long)interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
                 RealScreenCoord basepos;
                 basepos = MapDiagonalLength/2;
                 // Do the drawing
@@ -683,8 +683,8 @@ int draw_line_to_heart(struct PlayerInfo *player, long units_per_px, long zoom)
     // Now rotate the coordinates to receive minimap points
     RealScreenCoord mapos_x;
     RealScreenCoord mapos_y;
-    mapos_x = (zmpos_x * LbCosL(interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
-    mapos_y = (zmpos_y * LbCosL(interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL(interpolated_cam_rotation_angle_x)) >> 16;
+    mapos_x = (zmpos_x * LbCosL((long)interpolated_cam_rotation_angle_x) + zmpos_y * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
+    mapos_y = (zmpos_y * LbCosL((long)interpolated_cam_rotation_angle_x) - zmpos_x * LbSinL((long)interpolated_cam_rotation_angle_x)) >> 16;
     RealScreenCoord basepos;
     basepos = MapDiagonalLength/2;
     // Do the drawing
@@ -850,7 +850,7 @@ static void do_map_rotate_stuff(long relpos_x, long relpos_y, long *stl_x, long 
     const struct Camera *cam;
     cam = player->acamera;
     int angle;
-    angle = interpolated_cam_rotation_angle_x & ANGLE_MASK_4;
+    angle = (long)interpolated_cam_rotation_angle_x & ANGLE_MASK_4;
     int shift_x;
     int shift_y;
     shift_x = -LbSinL(angle);
@@ -1284,7 +1284,7 @@ void panel_map_draw_slabs(long x, long y, long units_per_px, long zoom)
     long shift_stl_y;
     {
         int angle;
-        angle = interpolated_cam_rotation_angle_x & ANGLE_MASK_4; //cam->rotation_angle_x
+        angle = (long)interpolated_cam_rotation_angle_x & ANGLE_MASK_4; //cam->rotation_angle_x
         shift_x = -LbSinL(angle) * zoom / 256;
         shift_y = LbCosL(angle) * zoom / 256;
         long current_minimap_x = (cam->mappos.x.stl.num << 16);
